// types.ts
export interface Vibe {
  id: string;
  name: string;
  icon: string;
}

export interface VoiceActor {
  id: string;
  name: string;
}

export interface VoiceVibe {
  id: string;
  name:string;
}

export type VideoModel = 'VEO 2.0';

export type AspectRatio = '1:1' | '9:16' | '16:9' | '4:3' | '3:4';